nsis: request admin execution level
The current state of our NSIS installer doesn't support per-user
installation and needs a major rework to do so, therefore the user
is required to have admin privileges to install VLC.

Still, RequestExecutionLevel was set to 'user' to be able to drop
privileges when launching vlc.exe at the end of the installation but
this was improved since (90496f23c2481bbaea21a586bc22631850751221) and
confirmed by the accesschk tool.

However requesting user execution level makes the unpacking occurs
twice: once with the user level and once after elevation.

Requesting admin permissions fixes the unpacking issue while still being
safe when launching VLC after the installation on XP and above.

Signed-off-by: Ludovic Fauvet <etix@videolan.org>