musicbrainz.lua: Use get_releaseid even if the album title is known
This renames `try_release` to `get_releaseid` and uses it to find the
MusicBrainz Identifier. That MBID is later used in the query passed to
try_query.

Not using the ASIN returned from the search server in the get_releaseid
call means one additional request is performed in the case that both the
artist and the album name are already known.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit f18236eee239fa8b57641ba82f20576ab244202f)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>