hotkeys: fullscreen state of the vout should have precedence over fullscreen state of the playlist
Double clicking to toggle to fullscreen mode does not trigger this code
path since 7b8083d228. The state of the playlist was therefore not in
sync with the vout. Pressing the fullscreen hotkey caused the state of
the playlist to be used to set the state of the vout, thus causing the bug.

Close #10433.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>
(cherry picked from commit 9f7bfd93e3a749a2912adda6380370310cd4dffa)
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>